requirements.txt
tensorflow<2.11
pydub
pytube
arabic-reshaper
fastapi
uvicorn
python-bidi
tensorflow-io==0.27.0